Rename main-menu to crud and move scenes to sub-folders
