+++ src/main/java/org/grupp2/sdpproject/GUI/customer/CustomerDashBoardScene.java
@@ -1,10 +1,10 @@
-package org.grupp2.sdpproject.GUI;
+package org.grupp2.sdpproject.GUI.customer;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
-import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
+import org.grupp2.sdpproject.GUI.SceneController;
 import org.grupp2.sdpproject.entities.User;
 
 public class CustomerDashBoardScene {
