+++ src/main/java/org/grupp2/sdpproject/GUI/customer/FilmsScene.java
@@ -1,10 +1,11 @@
-package org.grupp2.sdpproject.GUI;
+package org.grupp2.sdpproject.GUI.customer;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.layout.AnchorPane;
+import org.grupp2.sdpproject.GUI.SceneController;
 import org.grupp2.sdpproject.Utils.DAOManager;
 import org.grupp2.sdpproject.entities.Film;
