+++ src/main/resources/org/grupp2/sdpproject/add-special-features.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.layout.*?>
 
 <?import javafx.geometry.Insets?>
-<AnchorPane fx:id="root" prefHeight="300.0" prefWidth="300.0" stylesheets="@style.css" xmlns="http://javafx.com/javafx/17.0.12" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.grupp2.sdpproject.GUI.AddSpecialFeatures">
+<AnchorPane fx:id="root" prefHeight="300.0" prefWidth="300.0" stylesheets="@style.css" xmlns="http://javafx.com/javafx/17.0.12" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.grupp2.sdpproject.GUI.staff.AddSpecialFeatures">
     <children>
         <VBox alignment="TOP_CENTER" prefHeight="300.0" prefWidth="300.0">
             <children>
